Update ESMF SDK version and refactor imports
Bump esmf-sdk-version to 2.11.1 in pom.xml. Refactor imports in ModelUtils, ViolationFormatter, and ValidationUtils to use new locations for ProcessingViolation and InvalidSyntaxViolation, and update method signature in ModelUtils to use AspectModelFile.
